+++ HonjiMES/ViewModels/ReportModel.cs
@@ -69,7 +69,7 @@ namespace HonjiMES.Models
         public string ProductNo { get; set; }
         public string ProductName { get; set; }
         public int? Quantity { get; set; }
-        public int? Price { get; set; }
-        public int? Total { get; set; }
+        public decimal? Price { get; set; }
+        public decimal? Total { get; set; }
     }
 }
